requirements.txt
mkdocs-material
mkdocs-macros-plugin
mkdocs-redirects
mkdocstrings[python]
selenium
click
pyyaml
webdriver-manager
requests
pypdf
pillow
fastapi
uvicorn
markdown==3.6